add yml file and some

+++ database/migrations/2021_06_16_093353_create_hourly_ads_data_table.php
@@ -19,12 +19,12 @@ class CreateHourlyAdsDataTable extends Migration
             $table->datetime('time')->useCurrent();
             $table->integer('clicks');
             $table->integer('impressions');
-            $table->float('ctr', 4, 2);
-            $table->float('cost', 8, 2);
-            $table->float('cpc', 6, 2);
+            $table->float('ctr', 5, 2);
+            $table->integer('cost');
+            $table->integer('cpc');
             $table->integer('conversions');
             $table->float('conversions_rate', 5, 2);
-            $table->float('cost_per_conversion', 8, 2);
+            $table->integer('cost_per_conversion');
             $table->timestamps();
         });
     }
